+++ AirCondition/ViewModels/MapViewModel.swift
@@ -27,7 +27,7 @@ class MapViewModel: ViewModelType {
         self.output = Output()
         
         for i in 1..<10 {
-            self.output.devices.value.append(DeviceModel(serial: String(i), pm10: i, pm25: i*2, pm100: i*9, pressure: 1015, temperature: 16, humidity: 75, CO: Float(i) * 80.0, latitude: 52.0 + Double.random(in: 0.0..<1.0), longitude: 16.5 + Double.random(in: 0.0..<0.5) ))
+            self.output.devices.value.append(DeviceModel(serial: String(i), pm10: i, pm25: i*2, pm100: i*9, pressure: 1015, temperature: 16, humidity: 75, CO: Double(i) * 80.0, latitude: 52.0 + Double.random(in: 0.0..<1.0), longitude: 16.5 + Double.random(in: 0.0..<0.5) ))
         }
     }
